requirements.txt.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.8.0
python-dotenv==1.0.1
PyJWT==2.9.0
pytest==8.3.3